refactor: remove non necessary utils folder
