+++ src/writeErrorFile.js
